Enable skipping size and serializability profiling when always migrate/recompute is enabled (#403)
Enable skipping size and serializability profiling when always migrate/recompute is enabled for lower per-cell overheads as size/serializability results would not be used by the optimizer if those flags are on. Then, if an unserializable variable is sent for (incremental) checkpointing, it would be skipped (as it can be reconstructed on checkout).

Try statements were not added for non-incremental checkpointing as that option is not compatible with the always migrate/recompute flags.